// src/app/model/mpga/addmagasin.ts
export class AddMagasin {
    code_magasin:string;
    idMagasin:number;
    libelle: string;
    idZoneGeographique: number;
    nombreCaisse:number;
    numeroTelephone: string;
    fax: string;
    email: string;
    rue: string;
    idville: number;
    idResponsableZone: number;
    idResponsableMagasin:number;
    idAdjointResponsableMagasin:number;
    idWallet:number;
    affiliationMonetique: string;
    nombreCaissier:number;
    debutPlageAdresseIpCaisse: string;
    debutPlageAdresseIpTpe: string;
    adresseIpRacine: string;
    idGouvernorat:number;
    idCodePostal:number;
    constructor(){
    }
}